Update askstatus.sh and enhance update.sh script
Modified askstatus.sh to exit directly after case statement.
Enhanced update.sh with version comparison, root check, and
GitHub release fetching. Added version function, service stop,
and update execution.

+++ Tools/update.sh
@@ -1,5 +1,10 @@
 #!/usr/bin/env bash
 
+
+version() { 
+	echo "$@" | awk -F. '{ printf("%d%03d%03d%03d\n", $1,$2,$3,$4); }'; 
+}
+
 #Check if script is being run as root
 if [ "$(id -u)" != "0" ]; then
    echo "This script must be run as root" 1>&2
@@ -10,6 +15,21 @@ if [ ! $? = 0 ]; then
    exit 1
 else
    BASEDIR=${PWD}
+   latesturl="https://github.com/mickew/askstatus/releases/latest"
+   location=$(curl -s -I $latesturl | grep -i ^Location: | cut -d: -f2- | sed 's/^ *\(.*\).*/\1/')
+   ver="${location##*/}"
+   ver=$(echo "$ver" | sed 's/.\{1\}$//')
+
+   cd /var/www/backend/
+   oldver=$(/var/www/backend/Askstatus.Web.API --version)
+   cd ${BASEDIR}
+   oldver=$(echo "v$oldver")
+   if [ $(version $oldver) -ge $(version $ver) ]; then
+      echo "Version is up to date"
+      whiptail --title "Version is up to date" --msgbox "Version $ver => $oldver ." 8 78
+      exit
+   fi
+
    systemctl stop askstatusbackend.service
    sh ./getlatest.sh
